moved theme changing logic to theme.ts utility file
