+++ src/components/ThemeInitializer.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import { initializeTheme } from "@/utils/theme";
+import { useEffect } from "react";
+
+export default function ThemeInitializer() {
+  useEffect(() => {
+    initializeTheme();
+  }, []);
+
+  return <></>;
+}
